requirements.txt
aiogram==3.10.0
asyncpg==0.29.0
python-dotenv==1.0.1
cryptography==42.0.8
pytz==2025.2